+++ CHANGELOG.md
@@ -8,4 +8,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Import `lti_toolbox` library and tests from
+  [Ashley](https://github.com/openfun/ashley)
+
 [Unreleased]: https://github.com/openfun/django-lti-toolbox